+++ RestoRaterApp/ViewModels/UsersViewModel.swift
@@ -10,7 +10,11 @@ import CoreData
 
 final class UsersViewModel: ObservableObject {
     @Published var users: [User] = []
-
+    @Published var showingAlert = false
+    @Published var alertMessage = ""
+    @Published var userToDelete: User?
+    @Published var showingDeleteConfirmation = false
+    
     func fetchUsers(context: NSManagedObjectContext) {
         let request: NSFetchRequest<User> = User.createFetchRequest()
 
@@ -18,16 +22,31 @@ final class UsersViewModel: ObservableObject {
             users = try context.fetch(request)
         } catch {
             print("Error fetching restaurants: \(error)")
+            showingAlert = true
+            alertMessage = error.localizedDescription
         }
     }
 
-    func deleteUser(user: User, context: NSManagedObjectContext) {
-        context.delete(user)
+    func promptDelete(user: User) {
+        userToDelete = user
+        showingDeleteConfirmation = true
+    }
+    
+    func deleteUser(context: NSManagedObjectContext, completion: @escaping () -> Void) {
+        guard let user = userToDelete else {
+            showingAlert = true
+            alertMessage = "Something went wrong"
+            return
+        }
         
+        context.delete(user)
+        completion()
         do {
             try context.save()
         } catch {
             print("Error deleting restaurant: \(error)")
+            showingAlert = true
+            alertMessage = error.localizedDescription
         }
     }
 }
